// web/src/api/generated/model/listSplitPayments200PaymentsItemMethod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HouseApp API
 * API for HouseApp
 * OpenAPI spec version: 2.0.0
 */

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ListSplitPayments200PaymentsItemMethod = {  pix: 'pix',
  cash: 'cash',
  transfer: 'transfer',
  other: 'other',
  '': ,
  null: null,
} as const